+++ src/lib/symbols/marks.js
@@ -1,16 +1,14 @@
-
 /**
  * Replaces occurrences of a specific mark in a string.
  * 
  * Example: Company (tm) -> Company™
  *
  * @param {string} string - The input string where marks will be replaced.
- * @param {Object} locale - An object w/ locale-specific symbols.
  * @param {string} markPattern - The pattern for the mark (e.g., “tm” for trademark).
  * @param {string} replacementMark - The symbol to replace the pattern with (e.g., “™” for trademark).
+ * @param {Object} locale - An object w/ locale-specific symbols.
  * @returns {string} - The string with the specified marks replaced.
  */
-
 export function replaceMark(string, markPattern, replacementMark, locale) {
   return string.replace(
     new RegExp(
@@ -32,7 +30,6 @@ export function replaceMark(string, markPattern, replacementMark, locale) {
  * @param {Object} locale - An object w/ locale-specific symbols 
  * @returns {string} - The string with marks replaced.
  */
-
 export function fixMarks(string, locale) {
 
   string = replaceMark(string, "r", locale.registeredTrademark, locale);