+++ src/test/symbols/marks.test.js
@@ -72,7 +72,6 @@ function testMarks(testCase, markPattern, replacementMark) {
       });
     });
   });
-
 }